% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoMFA-package.R
\docType{package}
\name{autoMFA-package}
\alias{autoMFA}
\alias{autoMFA-package}
\title{autoMFA: Algorithms for Automatically Fitting MFA Models}
\description{
Provides methods for fitting the Mixture of Factor Analyzers (MFA) model automatically. 
    The MFA model is a mixture model where each sub-population is assumed to follow the Factor Analysis model. The Factor Analysis (FA) model is a latent variable model which assumes that observations are normally distributed, but imposes constraints on their covariance matrix. The MFA model contains two hyperparameters; g (the number of components in the mixture) and q (the number of factors in each component Factor Analysis model). Usually, the Expectation-Maximisation algorithm would be used to fit the MFA model, but this requires g and q to be known. This package treats g and q as unknowns and provides several methods which infer these values with as little input from the user as possible.
}
\author{
\strong{Maintainer}: John Davey \email{john.c.m.davey@gmail.com}

Other contributors:
\itemize{
  \item Sharon Lee [contributor]
  \item Garique Glonek [contributor]
  \item Suren Rathnayake [contributor]
  \item Geoff McLachlan [contributor]
  \item Albert Ali Salah [contributor]
  \item Heysem Kaya [contributor]
}

}
\keyword{internal}
