% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\docType{data}
\name{MFA_testdata}
\alias{MFA_testdata}
\title{Test dataset for the MFA model}
\format{
Data matrix with 720 observations of 3 variables.
}
\usage{
MFA_testdata
}
\description{
A 720 x 3 test dataset generated from a MFA model with 3 components, 1 factor for
each component. Uneven point distribution with large separation between
clusters relative to the component variance matrices.
}
\examples{
data(MFA_testdata)
plot(MFA_testdata[,1], MFA_testdata[,2])
}
\keyword{datasets}
